Update dataimport.m
In VHDR files, additional channels (such as ECG) are saved in the same file. Thus, with this change (line 72) we ensure that we don't consider it as a "brain electrode"

In the future, other "channels" recording physiological data (such as respiration) should be taken out. 

Note that I have not done it on (current) line 81 "%REMOVE PERIPHERAL CHANNELS" as this is specific for the VHDR format and the numbers of the channels can be wrongly considered (with our 127 electrodes + ECG, nbchan was equal to 128).